Update and rename subroutines/questions/prog_sub_12.cs to subroutines/questions/c-sharp/prog_sub_12.cs
